Rename `trade-utils.js` to `trade.js`
